+++ test/oap2/parameters/OAR022/fail-example.js
@@ -1,0 +1,33 @@
+module.exports = {
+  "swagger": "2.0",
+  "info": {
+    "version": "1.0.0",
+    "title": "Swagger Petstore"
+  },
+  "paths": {
+    "/orders": {
+      "get": {
+        "responses": { "206": { "description": "missing orderby" } }
+      }
+    },
+    "/customers": {
+      "get": {
+        "parameters": [],
+        "responses": { "206": { "description": "empty params" } }
+      }
+    },
+    "/users/{id}/orders": {
+      "get": {
+        "responses": { "206": { "description": "subcollection missing orderby" } }
+      }
+    },
+    "/products": {
+      "get": {
+        "parameters": [
+          { "name": "$limit", "in": "query", "type": "integer" }
+        ],
+        "responses": { "206": { "description": "wrong param" } }
+      }
+    }
+  }
+};