+++ test/oap2/parameters/oar025-limit-param.test.js
@@ -1,0 +1,23 @@
+const { linterForRule } = require('../../helpers/utils');
+
+let linter;
+
+const oar025fail = require('./OAR025/fail-example');
+const oar025ok = require('./OAR025/ok-example');
+
+beforeAll(async () => {
+  linter = await linterForRule('apiq:OAR025');
+  return linter;
+});
+
+test('apiq:OAR025 should find errors', () => {
+  return linter.run(oar025fail).then((results) => {
+    expect(results.length).toBe(3);
+  });
+});
+
+test('apiq:OAR025 should find no errors', () => {
+  return linter.run(oar025ok).then((results) => {
+    expect(results.length).toBe(0);
+  });
+});